quick_demos/distributed_tracing/service_b/requirements.txt
flask
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-flask==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-semantic-conventions==0.42b0
requests
